+++ src/main/webapp/resources/js/components/pipeline-launch/PipelineParameters.jsx
@@ -1,11 +1,10 @@
 import React, { useState } from "react";
-import { Collapse, List, Space, Typography } from "antd";
+import { Collapse, List, Radio, Typography, } from "antd";
 import { useLaunchDispatch, useLaunchState } from "./launch-context";
 import {
   DISPATCH_PARAMETER_CHANGE,
   DISPATCH_PARAMETERS_MODIFIED,
 } from "../pipeline/lauch-constants";
-import { IconCopy } from "../icons/Icons";
 
 const { Panel } = Collapse;
 const { Paragraph } = Typography;
@@ -31,25 +30,29 @@ export function PipelineParameters() {
     setSelected(Number(e.target.value));
   };
 
+  const duplicateParameters = (e) => {
+    e.stopPropagation();
+  };
+
   return (
     <>
       <Collapse expandIconPosition="right">
         {parameters.map((p) => (
           <Panel
             key={`parameter-${p.id}`}
-            extra={<IconCopy />}
             header={
-              <Space>
-                <input
+              <span onClick={(e) => e.stopPropagation()}>
+                <Radio
+                  style={{ height: 32 }}
                   type="radio"
                   name="para"
                   value={p.id}
                   onChange={updateSelected}
-                  onClick={(e) => e.stopPropagation()}
                   checked={selected === p.id}
-                />
-                {p.label}
-              </Space>
+                >
+                  {p.label}
+                </Radio>
+              </span>
             }
           >
             <List