+++ src/test/java/ca/corefacility/bioinformatics/irida/ria/integration/remoteapi/RemoteAPIDetailsPageIT.java
@@ -1,15 +1,16 @@
 package ca.corefacility.bioinformatics.irida.ria.integration.remoteapi;
 
+import org.junit.Test;
+
 import ca.corefacility.bioinformatics.irida.ria.integration.AbstractIridaUIITChromeDriver;
 import ca.corefacility.bioinformatics.irida.ria.integration.pages.LoginPage;
 import ca.corefacility.bioinformatics.irida.ria.integration.pages.remoteapi.RemoteAPIDetailsPage;
-import ca.corefacility.bioinformatics.irida.ria.integration.pages.remoteapi.RemoteAPIDetailsPage.ApiStatus;
+import ca.corefacility.bioinformatics.irida.ria.integration.pages.remoteapi.RemoteAPIsPage;
+
 import com.github.springtestdbunit.annotation.DatabaseSetup;
-import org.junit.Before;
-import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.assertFalse;
 
 /**
  * IT for the client details page
@@ -17,37 +18,33 @@ import static org.junit.Assert.assertTrue;
  */
 @DatabaseSetup("/ca/corefacility/bioinformatics/irida/ria/web/oauth/RemoteApisIT.xml")
 public class RemoteAPIDetailsPageIT extends AbstractIridaUIITChromeDriver {
-	private RemoteAPIDetailsPage page;
-
 	Long id = 1L;
 	String apiName = "a client";
 	String apiClient = "client";
-
-	@Before
-	public void setUpTest() {
-		LoginPage.loginAsManager(driver());
-
-		page = new RemoteAPIDetailsPage(driver(), id);
-	}
-
 	@Test
 	public void testVerifyClient() {
+		LoginPage.loginAsManager(driver());
+
+		RemoteAPIDetailsPage page = RemoteAPIDetailsPage.gotoDetailsPage(driver(), id);
 		String clientName = page.getClientName();
 		assertEquals(apiName, clientName);
 
 		String clientId = page.getClientId();
 		assertEquals(apiClient, clientId);
 
-		ApiStatus remoteApiStatus = page.getRemoteApiStatus();
-		assertEquals(ApiStatus.INVALID, remoteApiStatus);
+		assertFalse(page.isRemoteAPIConnected());
 
 	}
 
 	@Test
 	public void testDeleteClient() {
+		LoginPage.loginAsManager(driver());
+		RemoteAPIsPage listingPage = RemoteAPIsPage.goTo(driver());
+		int count = listingPage.remoteApisTableSize();
+		RemoteAPIDetailsPage page = RemoteAPIDetailsPage.gotoDetailsPage(driver(), id);
 		page.clickDeleteButton();
-		page.confirmDelete();
-		assertTrue(page.checkDeleteSuccess());
+		listingPage = RemoteAPIsPage.goTo(driver());
+		assertEquals(count - 1, listingPage.remoteApisTableSize());
 	}
 
 }